+++ components/src/lib/components/Button/Button.styles.ts
@@ -1,16 +1,5 @@
-import React from 'react';
 import styled from '@emotion/styled';
-
-type ButtonVariant = 'primary' | 'secondary';
-type ButtonSize = 'small' | 'medium' | 'large';
-
-export interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
-  children: React.ReactNode;
-  variant?: ButtonVariant;
-  size?: ButtonSize;
-  fullWidth?: boolean;
-  className?: string;
-}
+import { ButtonSize, ButtonVariant } from './Button.types';
 
 const StyledButton = styled.button<{
   variant: ButtonVariant;
@@ -80,16 +69,4 @@ const StyledButton = styled.button<{
   }
 `;
 
-export const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
-  ({ children, variant = 'primary', size = 'medium', fullWidth = false, className, ...props }, ref) => {
-    return (
-      <StyledButton ref={ref} variant={variant} size={size} fullWidth={fullWidth} className={className} {...props}>
-        {children}
-      </StyledButton>
-    );
-  },
-);
-
-Button.displayName = 'Button';
-
-export default Button;
+export { StyledButton };
