+++ components/src/lib/components/Button/Button.types.ts
@@ -1,0 +1,2 @@
+export type ButtonVariant = 'primary' | 'secondary';
+export type ButtonSize = 'small' | 'medium' | 'large';
