+++ components/src/lib/components/Input/Input.types.ts
@@ -1,0 +1,5 @@
+export interface InputProps extends React.InputHTMLAttributes<HTMLInputElement> {
+  fullWidth?: boolean;
+  error?: boolean;
+  className?: string;
+}
