+++ components/src/lib/components/Modal/CloseIcon.tsx
