+++ components/src/lib/index.ts
@@ -1,1 +1,1 @@
-export { Modal } from './Modal';
+export { Modal } from './components/Modal/Modal';
